requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.17
botocore==1.24.17
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.1
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
jmespath==0.10.0
mysqlclient==2.1.0
Naked==0.1.31
Pillow==9.0.1
pycrypto==2.6.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
